refactor: address PR review nits and privatize external_fetch internals #258
- tool_config_service: inline the single-use supports_url_attachments
  expression into the DialDeploymentTool(...) constructor (review nit).
- _mcp_tool: extract the dial_url file collection/validation loop out of
  _pre_process_params into a _collect_dial_url_files helper (review nit).
- Privatize internal external_fetch modules per protected-vs-public API
  feedback: external_fetch_settings, external_url_fetch_policy_resolver,
  host_pattern_match -> _-prefixed; imports and tests updated.
