+++ src/quickapp/shared/external_fetch/__init__.py
@@ -1,0 +1,3 @@
+from .external_url_fetcher import ExternalUrlFetcher
+
+__all__ = ["ExternalUrlFetcher"]
