+++ src/quickapp/shared/external_fetch/_external_fetch_settings.py
