+++ src/quickapp/shared/external_fetch/_external_url_fetch_policy_resolver.py
@@ -3,8 +3,8 @@ from typing import Literal
 from injector import inject
 
 from quickapp.config.application import ApplicationConfig
-from quickapp.shared.external_fetch.external_fetch_settings import ExternalFetchSettings
-from quickapp.shared.external_fetch.host_pattern_match import match_host
+from quickapp.shared.external_fetch._external_fetch_settings import ExternalFetchSettings
+from quickapp.shared.external_fetch._host_pattern_match import match_host
 
 PolicyReason = Literal["admin", "builder", "allowed"]
 HostReason = Literal["admin_allowlist", "builder_allowlist", "allowed"]
