+++ src/quickapp/shared/external_fetch/_host_pattern_match.py
