+++ src/tests/unit_tests/shared/test_external_url_fetch_policy_resolver.py
@@ -3,8 +3,8 @@ from unittest.mock import MagicMock
 import pytest
 
 from quickapp.config.application import ExternalUrlFetchConfig, Features, FileLoadingConfig
-from quickapp.shared.external_fetch.external_fetch_settings import ExternalFetchSettings
-from quickapp.shared.external_fetch.external_url_fetch_policy_resolver import (
+from quickapp.shared.external_fetch._external_fetch_settings import ExternalFetchSettings
+from quickapp.shared.external_fetch._external_url_fetch_policy_resolver import (
     ExternalUrlFetchPolicyResolver,
 )
