+++ src/tests/unit_tests/shared/test_external_url_fetcher.py
@@ -4,8 +4,8 @@ import httpx
 import pytest
 
 from quickapp.common.file_loading_size_limit_resolver import FileLoadingSizeLimitResolver
-from quickapp.shared.external_fetch.external_fetch_settings import ExternalFetchSettings
-from quickapp.shared.external_fetch.external_url_fetch_policy_resolver import (
+from quickapp.shared.external_fetch._external_fetch_settings import ExternalFetchSettings
+from quickapp.shared.external_fetch._external_url_fetch_policy_resolver import (
     ExternalUrlFetchPolicyResolver,
 )
 from quickapp.shared.external_fetch.external_url_fetcher import (
